requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
phonenumbers==8.13.53
PyJWT==2.10.1
python-decouple==3.8
razorpay==1.4.2
requests==2.32.3
sqlparse==0.5.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
